Atualiza arquivos do projeto
