--- .env
@@ -1,35 +1,0 @@
-# Configurações de bancos de dados
-BANCO_1=waybe
-BANCO_2=micro-waychef
-BANCO_3=sessao
-
-# Configurações de origens de banco de dados
-DB_1_IP=127.0.0.1
-DB_1_PORTA=3306
-DB_1_USER=root
-DB_1_PASS=root
-
-DB_2_IP=192.168.5.174
-DB_2_PORTA=3306
-DB_2_USER=root
-DB_2_PASS=root
-
-DB_3_IP=192.168.5.217
-DB_3_PORTA=3306
-DB_3_USER=root
-DB_3_PASS=root
-
-DB_4_IP=127.0.0.1
-DB_4_PORTA=3309
-DB_4_USER=root
-DB_4_PASS=root
-
-DB_5_IP=rasp.local
-DB_5_PORTA=3306
-DB_5_USER=root
-DB_5_PASS=root
-
-# Configurações de Java e Maven
-JAVA_HOME=/usr/lib/jvm/java-11-openjdk-amd64
-MAVEN_HOME=/opt/maven
-MAVEN_OPTS=-Xmx1024m
