+++ config.py
@@ -1,6 +1,9 @@
 import os
 import dotenv
 import re
+import tkinter as tk
+from tkinter import filedialog, messagebox
+import customtkinter as ctk
 from typing import Dict, Tuple, List
 
 # Caminho do arquivo .env
@@ -53,6 +56,7 @@ def get_db_origins() -> Dict[str, Tuple[str, str, str, str]]:
             f.write("DB_1_PORTA=3306\n")
             f.write("DB_1_USER=root\n")
             f.write("DB_1_PASS=root\n")
+            f.write("DB_1_NOME=Banco Local\n")
         
         # Recarrega o arquivo .env
         dotenv.load_dotenv(ENV_PATH, override=True)
@@ -85,6 +89,7 @@ def get_db_names() -> Dict[str, str]:
         # Adiciona um banco padrão ao arquivo .env
         with open(ENV_PATH, "a") as f:
             f.write("BANCO_1=waybe\n")
+            f.write("BANCO_1_NOME=Waybe\n")
         
         # Recarrega o arquivo .env
         dotenv.load_dotenv(ENV_PATH, override=True)
@@ -179,217 +184,322 @@ def get_next_id(prefix, env_path=".env"):
     ids = [int(id_str) for id_str in matches]
     return str(max(ids) + 1)
 
-# A interface gráfica de configuração só deve rodar se este arquivo for executado diretamente
+class ConfigApp:
+    def __init__(self, root):
+        self.root = root
+        self.root.title("Configuração do Sistema")
+        self.root.geometry("600x700")
+        
+        # Criar notebook (abas)
+        self.notebook = ctk.CTkTabview(self.root)
+        self.notebook.pack(fill="both", expand=True, padx=10, pady=10)
+        
+        # Criar abas
+        self.tab_java_maven = self.notebook.add("Java/Maven")
+        self.tab_origem = self.notebook.add("Origem do Banco")
+        self.tab_banco = self.notebook.add("Nome do Banco")
+        
+        # Configurar cada aba
+        self.setup_java_maven_tab()
+        self.setup_origem_tab()
+        self.setup_banco_tab()
+    
+    def setup_java_maven_tab(self):
+        # Título
+        ctk.CTkLabel(self.tab_java_maven, text="Configuração JAVA E MAVEN", font=("Arial", 16, "bold"), anchor="w").pack(anchor="w", padx=10, pady=5)
+        
+        # Frame para JAVA_HOME
+        java_frame = ctk.CTkFrame(self.tab_java_maven)
+        java_frame.pack(fill="x", padx=10, pady=5)
+        
+        ctk.CTkLabel(java_frame, text="JAVA_HOME:", width=100).pack(side="left", padx=5)
+        
+        self.entry_java = ctk.CTkEntry(java_frame, width=300)
+        self.entry_java.insert(0, os.getenv("JAVA_HOME", ""))
+        self.entry_java.pack(side="left", padx=5, fill="x", expand=True)
+        
+        def procurar_java_home():
+            diretorio = filedialog.askdirectory(title="Selecione o diretório JAVA_HOME")
+            if diretorio:
+                self.entry_java.delete(0, ctk.END)
+                self.entry_java.insert(0, diretorio)
+        
+        btn_procurar_java = ctk.CTkButton(java_frame, text="Procurar", width=80, command=procurar_java_home)
+        btn_procurar_java.pack(side="right", padx=5)
+        
+        # Frame para MAVEN_HOME
+        maven_frame = ctk.CTkFrame(self.tab_java_maven)
+        maven_frame.pack(fill="x", padx=10, pady=5)
+        
+        ctk.CTkLabel(maven_frame, text="MAVEN_HOME:", width=100).pack(side="left", padx=5)
+        
+        self.entry_maven = ctk.CTkEntry(maven_frame, width=300)
+        self.entry_maven.insert(0, os.getenv("MAVEN_HOME", ""))
+        self.entry_maven.pack(side="left", padx=5, fill="x", expand=True)
+        
+        def procurar_maven_home():
+            diretorio = filedialog.askdirectory(title="Selecione o diretório MAVEN_HOME")
+            if diretorio:
+                self.entry_maven.delete(0, ctk.END)
+                self.entry_maven.insert(0, diretorio)
+        
+        btn_procurar_maven = ctk.CTkButton(maven_frame, text="Procurar", width=80, command=procurar_maven_home)
+        btn_procurar_maven.pack(side="right", padx=5)
+        
+        # Botão para salvar
+        def salvar_java_maven():
+            # Atualiza o arquivo .env com os novos valores
+            java_home = self.entry_java.get()
+            maven_home = self.entry_maven.get()
+            
+            # Lê o conteúdo atual do arquivo .env
+            env_content = ""
+            if os.path.exists(ENV_PATH):
+                with open(ENV_PATH, "r") as f:
+                    env_content = f.read()
+            
+            # Remove as linhas existentes de JAVA_HOME e MAVEN_HOME
+            env_lines = env_content.splitlines()
+            env_lines = [line for line in env_lines if not line.startswith("JAVA_HOME=") and not line.startswith("MAVEN_HOME=")]
+            
+            # Adiciona as novas linhas
+            env_lines.append(f"JAVA_HOME={java_home}")
+            env_lines.append(f"MAVEN_HOME={maven_home}")
+            
+            # Escreve de volta no arquivo .env
+            with open(ENV_PATH, "w") as f:
+                f.write("\n".join(env_lines))
+            
+            # Recarrega as variáveis de ambiente
+            load_env_config()
+            
+            messagebox.showinfo("Salvo", "Configurações de JAVA e MAVEN salvas com sucesso.")
+        
+        btn_salvar = ctk.CTkButton(self.tab_java_maven, text="Salvar Configurações", command=salvar_java_maven)
+        btn_salvar.pack(pady=10)
+    
+    def setup_origem_tab(self):
+        # Título
+        ctk.CTkLabel(self.tab_origem, text="Configuração de Origem do banco", font=("Arial", 16, "bold"), anchor="w").pack(anchor="w", padx=10, pady=5)
+        
+        # Frame para lista de origens
+        self.origens_listbox_frame = ctk.CTkFrame(self.tab_origem)
+        self.origens_listbox_frame.pack(fill="both", expand=True, padx=10, pady=5)
+        
+        ctk.CTkLabel(self.origens_listbox_frame, text="Origens cadastradas:", anchor="w").pack(anchor="w", padx=5, pady=5)
+        
+        # Listbox para origens
+        self.origens_listbox = tk.Listbox(self.origens_listbox_frame, bg="#2b2b2b", fg="white", font=("Arial", 12), height=10)
+        self.origens_listbox.pack(fill="both", expand=True, padx=5, pady=5)
+        
+        # Atualiza a lista de origens
+        self.atualizar_lista_origens()
+        
+        # Frame para nova origem
+        nova_origem_frame = ctk.CTkFrame(self.tab_origem)
+        nova_origem_frame.pack(fill="x", padx=10, pady=5)
+        
+        # Campos para nova origem
+        ctk.CTkLabel(nova_origem_frame, text="Nova Origem:", anchor="w").pack(anchor="w", padx=5, pady=5)
+        
+        # IP
+        ip_frame = ctk.CTkFrame(nova_origem_frame)
+        ip_frame.pack(fill="x", padx=5, pady=2)
+        ctk.CTkLabel(ip_frame, text="IP:", width=80).pack(side="left", padx=5)
+        self.entry_ip = ctk.CTkEntry(ip_frame)
+        self.entry_ip.pack(side="left", fill="x", expand=True, padx=5)
+        
+        # Porta
+        porta_frame = ctk.CTkFrame(nova_origem_frame)
+        porta_frame.pack(fill="x", padx=5, pady=2)
+        ctk.CTkLabel(porta_frame, text="Porta:", width=80).pack(side="left", padx=5)
+        self.entry_porta = ctk.CTkEntry(porta_frame)
+        self.entry_porta.insert(0, "3306")
+        self.entry_porta.pack(side="left", fill="x", expand=True, padx=5)
+        
+        # Usuário
+        usuario_frame = ctk.CTkFrame(nova_origem_frame)
+        usuario_frame.pack(fill="x", padx=5, pady=2)
+        ctk.CTkLabel(usuario_frame, text="Usuário:", width=80).pack(side="left", padx=5)
+        self.entry_usuario = ctk.CTkEntry(usuario_frame)
+        self.entry_usuario.insert(0, "root")
+        self.entry_usuario.pack(side="left", fill="x", expand=True, padx=5)
+        
+        # Senha
+        senha_frame = ctk.CTkFrame(nova_origem_frame)
+        senha_frame.pack(fill="x", padx=5, pady=2)
+        ctk.CTkLabel(senha_frame, text="Senha:", width=80).pack(side="left", padx=5)
+        self.entry_senha = ctk.CTkEntry(senha_frame, show="*")
+        self.entry_senha.insert(0, "root")
+        self.entry_senha.pack(side="left", fill="x", expand=True, padx=5)
+        
+        # Nome amigável
+        nome_frame = ctk.CTkFrame(nova_origem_frame)
+        nome_frame.pack(fill="x", padx=5, pady=2)
+        ctk.CTkLabel(nome_frame, text="Nome:", width=80).pack(side="left", padx=5)
+        self.entry_nome_origem = ctk.CTkEntry(nome_frame)
+        self.entry_nome_origem.pack(side="left", fill="x", expand=True, padx=5)
+        
+        # Botão para adicionar
+        def adicionar_origem():
+            # Gera ID automaticamente
+            next_id = get_next_id("DB")
+            prefix = f"DB_{next_id}"
+            
+            # Verifica se os campos obrigatórios estão preenchidos
+            if not self.entry_ip.get():
+                messagebox.showerror("Erro", "O campo IP é obrigatório.")
+                return
+            
+            # Obtém os valores dos campos
+            ip = self.entry_ip.get()
+            porta = self.entry_porta.get() or '3306'
+            usuario = self.entry_usuario.get() or 'root'
+            senha = self.entry_senha.get() or 'root'
+            
+            # Cria um nome amigável para a origem
+            nome_amigavel = self.entry_nome_origem.get() or f"Banco {ip}:{porta}"
+            
+            # Adiciona ao arquivo .env
+            with open(ENV_PATH, "a") as f:
+                f.write(f"{prefix}_IP={ip}\n")
+                f.write(f"{prefix}_PORTA={porta}\n")
+                f.write(f"{prefix}_USER={usuario}\n")
+                f.write(f"{prefix}_PASS={senha}\n")
+                f.write(f"{prefix}_NOME={nome_amigavel}\n")
+            
+            # Recarrega as variáveis de ambiente
+            load_env_config()
+            
+            messagebox.showinfo("Salvo", f"Origem {nome_amigavel} cadastrada com sucesso.")
+            
+            # Limpa os campos após o cadastro
+            self.entry_ip.delete(0, ctk.END)
+            self.entry_porta.delete(0, ctk.END)
+            self.entry_porta.insert(0, "3306")
+            self.entry_usuario.delete(0, ctk.END)
+            self.entry_usuario.insert(0, "root")
+            self.entry_senha.delete(0, ctk.END)
+            self.entry_senha.insert(0, "root")
+            self.entry_nome_origem.delete(0, ctk.END)
+            
+            # Atualiza a lista de origens
+            self.atualizar_lista_origens()
+        
+        btn_adicionar = ctk.CTkButton(nova_origem_frame, text="Adicionar Origem", command=adicionar_origem)
+        btn_adicionar.pack(pady=10)
+    
+    def atualizar_lista_origens(self):
+        # Limpa a lista atual
+        self.origens_listbox.delete(0, tk.END)
+        
+        # Obtém as origens e nomes de exibição
+        origens = get_db_origins()
+        display_names = get_origin_display_names()
+        
+        # Adiciona cada origem à lista
+        for db_id, (ip, porta, usuario, _) in origens.items():
+            nome = display_names.get(db_id, f"Banco {ip}:{porta}")
+            self.origens_listbox.insert(tk.END, f"{nome} ({ip}:{porta})")
+    
+    def setup_banco_tab(self):
+        # Título
+        ctk.CTkLabel(self.tab_banco, text="Cadastro de Nome de Banco", font=("Arial", 16, "bold"), anchor="w").pack(anchor="w", padx=10, pady=5)
+        
+        # Frame para lista de bancos
+        self.bancos_listbox_frame = ctk.CTkFrame(self.tab_banco)
+        self.bancos_listbox_frame.pack(fill="both", expand=True, padx=10, pady=5)
+        
+        ctk.CTkLabel(self.bancos_listbox_frame, text="Bancos cadastrados:", anchor="w").pack(anchor="w", padx=5, pady=5)
+        
+        # Listbox para bancos
+        self.bancos_listbox = tk.Listbox(self.bancos_listbox_frame, bg="#2b2b2b", fg="white", font=("Arial", 12), height=10)
+        self.bancos_listbox.pack(fill="both", expand=True, padx=5, pady=5)
+        
+        # Atualiza a lista de bancos
+        self.atualizar_lista_bancos()
+        
+        # Frame para novo banco
+        novo_banco_frame = ctk.CTkFrame(self.tab_banco)
+        novo_banco_frame.pack(fill="x", padx=10, pady=5)
+        
+        # Campos para novo banco
+        ctk.CTkLabel(novo_banco_frame, text="Novo Banco:", anchor="w").pack(anchor="w", padx=5, pady=5)
+        
+        # Nome do banco
+        nome_banco_frame = ctk.CTkFrame(novo_banco_frame)
+        nome_banco_frame.pack(fill="x", padx=5, pady=2)
+        ctk.CTkLabel(nome_banco_frame, text="Nome:", width=80).pack(side="left", padx=5)
+        self.entry_nome_banco = ctk.CTkEntry(nome_banco_frame)
+        self.entry_nome_banco.pack(side="left", fill="x", expand=True, padx=5)
+        
+        # Nome amigável
+        nome_amigavel_frame = ctk.CTkFrame(novo_banco_frame)
+        nome_amigavel_frame.pack(fill="x", padx=5, pady=2)
+        ctk.CTkLabel(nome_amigavel_frame, text="Exibição:", width=80).pack(side="left", padx=5)
+        self.entry_nome_amigavel = ctk.CTkEntry(nome_amigavel_frame)
+        self.entry_nome_amigavel.pack(side="left", fill="x", expand=True, padx=5)
+        
+        # Botão para adicionar
+        def adicionar_banco():
+            # Gera ID automaticamente
+            next_id = get_next_id("BANCO")
+            
+            # Verifica se o campo está preenchido
+            if not self.entry_nome_banco.get():
+                messagebox.showerror("Erro", "O nome do banco é obrigatório.")
+                return
+            
+            # Obtém o nome do banco
+            nome_banco = self.entry_nome_banco.get()
+            
+            # Cria um nome amigável para o banco
+            nome_amigavel = self.entry_nome_amigavel.get() or nome_banco.capitalize()
+            
+            # Adiciona ao arquivo .env
+            with open(ENV_PATH, "a") as f:
+                f.write(f"BANCO_{next_id}={nome_banco}\n")
+                f.write(f"BANCO_{next_id}_NOME={nome_amigavel}\n")
+            
+            # Recarrega as variáveis de ambiente
+            load_env_config()
+            
+            messagebox.showinfo("Salvo", f"Banco {nome_amigavel} cadastrado com sucesso.")
+            
+            # Limpa os campos após o cadastro
+            self.entry_nome_banco.delete(0, ctk.END)
+            self.entry_nome_amigavel.delete(0, ctk.END)
+            
+            # Atualiza a lista de bancos
+            self.atualizar_lista_bancos()
+        
+        btn_adicionar = ctk.CTkButton(novo_banco_frame, text="Adicionar Banco", command=adicionar_banco)
+        btn_adicionar.pack(pady=10)
+    
+    def atualizar_lista_bancos(self):
+        # Limpa a lista atual
+        self.bancos_listbox.delete(0, tk.END)
+        
+        # Obtém os bancos e nomes de exibição
+        bancos = get_db_names()
+        display_names = get_db_display_names()
+        
+        # Adiciona cada banco à lista
+        for db_id, nome in bancos.items():
+            display_name = display_names.get(db_id, nome.capitalize())
+            self.bancos_listbox.insert(tk.END, f"{display_name} ({nome})")
+
 if __name__ == "__main__":
-    import customtkinter as ctk
-    from tkinter import messagebox, filedialog
-
-    # Carrega variáveis do .env
+    # Carrega as configurações do arquivo .env
     load_env_config()
-
+    
     # UI Principal
     ctk.set_appearance_mode("dark")
     ctk.set_default_color_theme("blue")
     app = ctk.CTk()
-    app.title("Configurador Maven e Banco de Dados")
-    app.geometry("800x600")
-
-    # === Frame JAVA/MAVEN ===
-    frame_java = ctk.CTkFrame(app)
-    frame_java.pack(pady=10, fill="x", padx=20)
-    ctk.CTkLabel(frame_java, text="Configuração JAVA E MAVEN", font=("Arial", 18, "bold"), anchor="w").pack(anchor="w", padx=10, pady=5)
-
-    # Frame para JAVA_HOME com botão de procurar
-    java_frame = ctk.CTkFrame(frame_java)
-    java_frame.pack(fill="x", padx=10, pady=5)
     
-    entry_java = ctk.CTkEntry(java_frame, placeholder_text="JAVA_HOME", width=500)
-    entry_java.insert(0, os.getenv("JAVA_HOME", ""))
-    entry_java.pack(side="left", padx=(0, 10), fill="x", expand=True)
+    # Inicializa a aplicação
+    config_app = ConfigApp(app)
     
-    def procurar_java_home():
-        diretorio = filedialog.askdirectory(title="Selecione o diretório JAVA_HOME")
-        if diretorio:
-            entry_java.delete(0, ctk.END)
-            entry_java.insert(0, diretorio)
-    
-    btn_procurar_java = ctk.CTkButton(java_frame, text="Procurar", command=procurar_java_home)
-    btn_procurar_java.pack(side="right")
-
-    # Frame para MAVEN_HOME com botão de procurar
-    maven_frame = ctk.CTkFrame(frame_java)
-    maven_frame.pack(fill="x", padx=10, pady=5)
-    
-    entry_maven = ctk.CTkEntry(maven_frame, placeholder_text="MAVEN_HOME", width=500)
-    entry_maven.insert(0, os.getenv("MAVEN_HOME", ""))
-    entry_maven.pack(side="left", padx=(0, 10), fill="x", expand=True)
-    
-    def procurar_maven_home():
-        diretorio = filedialog.askdirectory(title="Selecione o diretório MAVEN_HOME")
-        if diretorio:
-            entry_maven.delete(0, ctk.END)
-            entry_maven.insert(0, diretorio)
-    
-    btn_procurar_maven = ctk.CTkButton(maven_frame, text="Procurar", command=procurar_maven_home)
-    btn_procurar_maven.pack(side="right")
-
-    # === Frame para adicionar nova origem de banco ===
-    frame_origem = ctk.CTkFrame(app)
-    frame_origem.pack(pady=10, fill="x", padx=20)
-    ctk.CTkLabel(frame_origem, text="Configuração de Origem do banco", font=("Arial", 18, "bold"), anchor="w").pack(anchor="w", padx=10, pady=5)
-
-    # Exibe as origens existentes
-    origens_frame = ctk.CTkFrame(frame_origem)
-    origens_frame.pack(fill="x", padx=10, pady=5)
-    
-    origens = get_db_origins()
-    display_names = get_origin_display_names()
-    
-    if origens:
-        ctk.CTkLabel(origens_frame, text="Origens cadastradas:").pack(anchor="w", padx=5, pady=5)
-        for db_id, (ip, porta, usuario, _) in origens.items():
-            nome = display_names.get(db_id, f"Banco {ip}:{porta}")
-            ctk.CTkLabel(origens_frame, text=f"ID: {db_id} - {nome} ({ip}:{porta}, usuário: {usuario})").pack(anchor="w", padx=20, pady=2)
-    
-    # Campos para nova origem
-    entry_ip = ctk.CTkEntry(frame_origem, placeholder_text="IP")
-    entry_porta = ctk.CTkEntry(frame_origem, placeholder_text="Porta")
-    entry_usuario = ctk.CTkEntry(frame_origem, placeholder_text="Usuário")
-    entry_senha = ctk.CTkEntry(frame_origem, placeholder_text="Senha", show="*")
-
-    for entry in [entry_ip, entry_porta, entry_usuario, entry_senha]:
-        entry.pack(padx=10, pady=2)
-
-    def adicionar_origem():
-        # Gera ID automaticamente
-        next_id = get_next_id("DB")
-        prefix = f"DB_{next_id}"
-        
-        # Verifica se os campos obrigatórios estão preenchidos
-        if not entry_ip.get():
-            messagebox.showerror("Erro", "O campo IP é obrigatório.")
-            return
-        
-        # Obtém os valores dos campos
-        ip = entry_ip.get()
-        porta = entry_porta.get() or '3306'
-        usuario = entry_usuario.get() or 'root'
-        senha = entry_senha.get() or 'root'
-        
-        # Cria um nome amigável para a origem
-        nome_amigavel = f"Banco {ip}:{porta}"
-        
-        # Adiciona ao arquivo .env
-        with open(ENV_PATH, "a") as f:
-            f.write(f"{prefix}_IP={ip}\n")
-            f.write(f"{prefix}_PORTA={porta}\n")
-            f.write(f"{prefix}_USER={usuario}\n")
-            f.write(f"{prefix}_PASS={senha}\n")
-            f.write(f"{prefix}_NOME={nome_amigavel}\n")
-        
-        # Recarrega as variáveis de ambiente
-        load_env_config()
-        
-        messagebox.showinfo("Salvo", f"Origem {next_id} cadastrada com sucesso.")
-        
-        # Limpa os campos após o cadastro
-        for entry in [entry_ip, entry_porta, entry_usuario, entry_senha]:
-            entry.delete(0, ctk.END)
-            
-        # Reinicia a aplicação para atualizar a lista de origens
-        app.destroy()
-        os.system(f"python {__file__}")
-
-    btn_add_origem = ctk.CTkButton(frame_origem, text="Cadastrar Origem", command=adicionar_origem)
-    btn_add_origem.pack(padx=10, pady=5)
-
-    # === Frame para adicionar nome de banco ===
-    frame_banco = ctk.CTkFrame(app)
-    frame_banco.pack(pady=10, fill="x", padx=20)
-    ctk.CTkLabel(frame_banco, text="Cadastro de Nome de Banco", font=("Arial", 18, "bold"), anchor="w").pack(anchor="w", padx=10, pady=5)
-
-    # Exibe os bancos existentes
-    bancos_frame = ctk.CTkFrame(frame_banco)
-    bancos_frame.pack(fill="x", padx=10, pady=5)
-    
-    bancos = get_db_names()
-    display_names = get_db_display_names()
-    
-    if bancos:
-        ctk.CTkLabel(bancos_frame, text="Bancos cadastrados:").pack(anchor="w", padx=5, pady=5)
-        for db_id, nome in bancos.items():
-            display_name = display_names.get(db_id, nome)
-            ctk.CTkLabel(bancos_frame, text=f"ID: {db_id} - {display_name}").pack(anchor="w", padx=20, pady=2)
-    
-    # Campo para novo banco
-    entry_banco_nome = ctk.CTkEntry(frame_banco, placeholder_text="Nome do Banco")
-    entry_banco_nome.pack(padx=10, pady=2)
-
-    def adicionar_banco():
-        # Gera ID automaticamente
-        next_id = get_next_id("BANCO")
-        
-        # Verifica se o campo está preenchido
-        if not entry_banco_nome.get():
-            messagebox.showerror("Erro", "O nome do banco é obrigatório.")
-            return
-        
-        # Obtém o nome do banco
-        nome_banco = entry_banco_nome.get()
-        
-        # Cria um nome amigável para o banco (primeira letra maiúscula)
-        nome_amigavel = nome_banco.capitalize()
-        
-        # Adiciona ao arquivo .env
-        with open(ENV_PATH, "a") as f:
-            f.write(f"BANCO_{next_id}={nome_banco}\n")
-            f.write(f"BANCO_{next_id}_NOME={nome_amigavel}\n")
-        
-        # Recarrega as variáveis de ambiente
-        load_env_config()
-        
-        messagebox.showinfo("Salvo", f"Banco {next_id} cadastrado com sucesso.")
-        
-        # Limpa o campo após o cadastro
-        entry_banco_nome.delete(0, ctk.END)
-        
-        # Reinicia a aplicação para atualizar a lista de bancos
-        app.destroy()
-        os.system(f"python {__file__}")
-
-    btn_add_banco = ctk.CTkButton(frame_banco, text="Cadastrar Banco", command=adicionar_banco)
-    btn_add_banco.pack(padx=10, pady=5)
-
-    # === Botão para salvar JAVA/MAVEN ===
-    def salvar_java_maven():
-        # Atualiza o arquivo .env com os novos valores
-        java_home = entry_java.get()
-        maven_home = entry_maven.get()
-        
-        # Lê o conteúdo atual do arquivo
-        env_content = ""
-        if os.path.exists(ENV_PATH):
-            with open(ENV_PATH, "r") as f:
-                env_content = f.read()
-        
-        # Remove as linhas existentes de JAVA_HOME e MAVEN_HOME
-        lines = env_content.splitlines()
-        new_lines = [line for line in lines if not line.startswith("JAVA_HOME=") and not line.startswith("MAVEN_HOME=")]
-        
-        # Adiciona as novas linhas
-        new_lines.append(f"JAVA_HOME={java_home}")
-        new_lines.append(f"MAVEN_HOME={maven_home}")
-        
-        # Escreve de volta no arquivo
-        with open(ENV_PATH, "w") as f:
-            f.write("\n".join(new_lines) + "\n")
-        
-        # Recarrega as variáveis de ambiente
-        load_env_config()
-        
-        messagebox.showinfo("Salvo", "Caminhos JAVA_HOME e MAVEN_HOME salvos no .env")
-
-    btn_salvar_geral = ctk.CTkButton(app, text="Salvar JAVA/MAVEN", command=salvar_java_maven)
-    btn_salvar_geral.pack(pady=10)
-
+    # Inicia o loop principal
     app.mainloop()
